requirements.txt
Django==4.2.10
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.1.0
djangorestframework==3.14.0
Pillow==10.2.0
django-ckeditor==6.7.1
cloudinary==1.37.0 